+++ tests/test_middleware.py
@@ -1,5 +1,6 @@
 from __future__ import annotations
 
+from django.core.exceptions import MiddlewareNotUsed
 from django.http import HttpRequest, HttpResponse, StreamingHttpResponse
 from django.test import RequestFactory, SimpleTestCase, override_settings
 
@@ -18,6 +19,14 @@ class BrowserReloadMiddlewareTests(SimpleTestCase):
             return self.response
 
         self.middleware = BrowserReloadMiddleware(get_response)
+
+    @override_settings(DEBUG=False)
+    def test_not_debug_instantiation(self):
+        def get_response(request: HttpRequest) -> HttpResponse:
+            return self.response
+
+        with self.assertRaises(MiddlewareNotUsed):
+            BrowserReloadMiddleware(get_response)
 
     @override_settings(DEBUG=False)
     def test_not_debug(self):